feat!: Use the alias edx/brand package.
This repo was not using the `edx/brand` package like most other frontend
repos.  It was instead hardcoded to using the edx.org branding.  This
change updates the repo to default to using the `edx/brand` package
aliased to the `oppenedx/brand-openedx` package by default.

BREAKING CHANGE:  If you wery relying on defults set in the
edx/brand-edx.org package.  You will need to alias the `@edx/brand`
package to `@edx/brand-edx.org` or your own fork of that package.